Use Throttling Netty Write Handler on HTTP Path
This adds the throttling write logic we have on the transport path to the HTTP path as well by
extracting the logic to a separate handler and using it in both pipelines.
This significantly reduces the heap consumption of writing large responses on coordinating nodes
when TLS is used by the same mechanism it does so on the transport path.
